pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Offline-first RAG chatbot starter stack for Raspberry Pi"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "chromadb",
    "ollama",
    "flask",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
upgrade = ["fastapi", "uvicorn[standard]", "pymupdf"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
